demo/requirements.txt
flask>=2.0.0
pillow>=9.0.0
requests>=2.25.0
pandas>=1.3.0
numpy>=1.21.0 